hastaExamFailNull

+++ examen02/app/src/main/java/com/example/examenvickybonilla/BFarmacia.kt
@@ -3,7 +3,7 @@ package com.example.examenvickybonilla
 import kotlin.collections.ArrayList
 
 class BFarmacia(
-    var idF:Int?,
+    var idF:Long?,
     var nombreF:String?,
     var meds: ArrayList<BMedicamento> = arrayListOf<BMedicamento>()
 )
